requirements.txt
GitPython
re
signal
sys
time
json